backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
python-dateutil==2.9.0.post0
pydantic==2.9.2
apscheduler==3.10.4
ics==0.7.2
httpx==0.27.2
python-dotenv==1.0.1
pywebpush==1.14.1
pyyaml==6.0.2
# >>> AUTH START
python-jose==3.3.0
passlib[bcrypt]==1.7.4
# <<< AUTH END
# >>> DFE START
Jinja2==3.1.4
# <<< DFE END
